requirements.txt
Django==1.8.7
MySQL-python==1.2.5
Pillow==3.0.0
Pygments==2.0.2
argparse==1.2.1
blessings==1.6
bootstrap-admin==0.3.6
bpython==0.14.2
curtsies==0.1.23
django-audit-log==0.7.0
django-localflavor==1.2
django-model-utils==2.4
django-smart-selects==1.2.0
greenlet==0.4.9
requests==2.9.1
six==1.10.0
sorl-thumbnail==12.3
wsgiref==0.1.2
